neuroscope: compute traceview ordinateMin in long arithmetic (hardening)
TraceView::updateWindow() derives the world lower bound as the negation
of a sum of layout dimensions:

    ordinateMin = -(2*borderY + nbShownchannels*traceVspace + ...);

Today every term is a signed int member, so the unary minus yields a
correct negative int that widens into the long ordinateMin, and the
height ordinateMax - ordinateMin is positive.  This is behaviour we want
to keep, not a bug being fixed.

The hardening guards against the failure mode that bit ErrorMatrixView,
where a single member typed 'uint' made the whole summand unsigned and
the unary minus wrapped (-6390 -> 4294960906), producing a negative
height that silently disabled zoom/pan while QPoint truncation kept the
render correct and masked it.  If any of these layout members is later
changed to an unsigned type, an int-typed sum here would wrap the same
way.

Force long arithmetic for the four ordinateMin negations (single/multi
column x raster/no-raster) by widening the leading term and each
data-scaling product, mirroring the klusters ErrorMatrixView fix.  The
values are unchanged for all realistic inputs (the operands are small,
non-negative pixel/count quantities); only the arithmetic type changes,
so the negation cannot wrap regardless of future member-type edits.
abscissaMax is left as-is: it has no unary minus and cannot wrap.

Built clean (neuroscope relinks), opens 3/3 headless.

+++ src/neuroscope/src/traceview.cpp
@@ -1201,7 +1201,10 @@ void TraceView::updateWindow(){
         if (raster && nbClusters != 0){
             int rasterTotalHeight = YTracesRasterSeparator + nbClusters * rasterHeight + (nbClusters - 1) * YRasterSpace;
 
-            ordinateMin = -(borderY + borderY/2 + (nbGps - 1) * YGroupSpace + nbShownchannels * traceVspace + nbYspaces * Yspace
+            ordinateMin = -(static_cast<long>(borderY) + borderY/2
+                            + static_cast<long>(nbGps - 1) * YGroupSpace
+                            + static_cast<long>(nbShownchannels) * traceVspace
+                            + static_cast<long>(nbYspaces) * Yspace
                             + rasterTotalHeight);
 
             Y0 = borderY/2 + (nbGps - 1) * YGroupSpace + nbShownchannels * traceVspace + nbYspaces * Yspace - (traceVspace/2)
@@ -1210,7 +1213,10 @@ void TraceView::updateWindow(){
             Y0Raster = borderY/2 + rasterTotalHeight - YTracesRasterSeparator;
         }
         else{
-            ordinateMin = -(2 * borderY + (nbGps - 1) * YGroupSpace + nbShownchannels * traceVspace + nbYspaces * Yspace);
+            ordinateMin = -(2L * borderY
+                            + static_cast<long>(nbGps - 1) * YGroupSpace
+                            + static_cast<long>(nbShownchannels) * traceVspace
+                            + static_cast<long>(nbYspaces) * Yspace);
             Y0 = borderY + (nbGps - 1) * YGroupSpace + nbShownchannels * traceVspace + nbYspaces * Yspace - (traceVspace/2);
         }
     }
@@ -1276,7 +1282,9 @@ void TraceView::updateWindow(){
 
             int rasterTotalHeight = YTracesRasterSeparator + maxNbClusters * rasterHeight + (maxNbClusters - 1) * YRasterSpace;
 
-            ordinateMin = -(borderY + borderY/2 + maxNbChannels * traceVspace + (maxNbChannels - 1) * Yspace
+            ordinateMin = -(static_cast<long>(borderY) + borderY/2
+                            + static_cast<long>(maxNbChannels) * traceVspace
+                            + static_cast<long>(maxNbChannels - 1) * Yspace
                             + rasterTotalHeight);
             Y0 = borderY/2 + maxNbChannels * traceVspace + (maxNbChannels - 1) * Yspace - (traceVspace/2)
                     + rasterTotalHeight;
@@ -1284,7 +1292,9 @@ void TraceView::updateWindow(){
             Y0Raster = borderY/2 + rasterTotalHeight - YTracesRasterSeparator;
         }
         else{
-            ordinateMin = -(2 * borderY + maxNbChannels * traceVspace + (maxNbChannels - 1) * Yspace);
+            ordinateMin = -(2L * borderY
+                            + static_cast<long>(maxNbChannels) * traceVspace
+                            + static_cast<long>(maxNbChannels - 1) * Yspace);
             Y0 = borderY + maxNbChannels * traceVspace + (maxNbChannels - 1) * Yspace - (traceVspace/2);
         }
     }
